+++ src/interface/TWStellarSigner.cpp
@@ -4,20 +4,20 @@
 // terms governing use, modification, and redistribution, is contained in the
 // file LICENSE at the root of the source code distribution tree.
 
+#include "../PrivateKey.h"
+#include "../proto/Stellar.pb.h"
+#include "../Stellar/Address.h"
+#include "../Stellar/Signer.h"
+
 #include <TrustWalletCore/TWStellarSigner.h>
 #include <TrustWalletCore/TWPrivateKey.h>
-#include <PrivateKey.h>
-
-#include "../Stellar/Signer.h"
-#include "../Stellar/Address.h"
-#include "../proto/Stellar.pb.h"
 
 using namespace TW;
 using namespace TW::Stellar;
 
 TW_Stellar_Proto_SigningOutput TWStellarSignerSign(TW_Stellar_Proto_SigningInput data) {
     Proto::SigningInput input;
-    input.ParseFromArray(TWDataBytes(data), TWDataSize(data));
+    input.ParseFromArray(TWDataBytes(data), static_cast<int>(TWDataSize(data)));
 
     const auto signer = Signer(input);
     const auto output = signer.sign();
